+++ src/services/asset-inventory/service-account/service-account-detail/modules/ServiceAccountProjectDetail.vue
